Move all nn tests to single file
